% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhhrvserver.R
\name{HrvServer}
\alias{HrvServer}
\title{Title 预览下载}
\usage{
HrvServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值

返回值

返回值
}
\description{
Title 预览下载
}
\examples{
downloadserver()
Title 下载日志

downloadserver()
Title 后台处理总函数

HrvServer()
}
